+++ Config.lua
@@ -1105,7 +1105,6 @@ DF.PartyDefaults = {
     directBuffFilterRaidInCombat = true,      -- RAID_IN_COMBAT filter
     directBuffFilterCancelable = false,       -- CANCELABLE filter
     directBuffFilterNotCancelable = false,    -- NOT_CANCELABLE filter
-    directBuffFilterImportant = true,         -- IMPORTANT filter (12.0.1)
     directBuffFilterBigDefensive = true,      -- BIG_DEFENSIVE filter (12.0.1)
     directBuffFilterExternalDefensive = true, -- EXTERNAL_DEFENSIVE filter (12.0.0)
     directBuffSortOrder = "TIME",             -- "DEFAULT" / "TIME" / "NAME"
@@ -1115,7 +1114,6 @@ DF.PartyDefaults = {
     directDebuffFilterRaid = true,            -- RAID filter
     directDebuffFilterRaidInCombat = true,    -- RAID_IN_COMBAT filter
     directDebuffFilterCrowdControl = true,    -- CROWD_CONTROL filter
-    directDebuffFilterImportant = true,       -- IMPORTANT filter (12.0.1)
     directDebuffDispellableMode = "PLAYER",  -- "PLAYER" (RAID_PLAYER_DISPELLABLE) or "ALL" (dispelName ~= nil)
     directDebuffSortOrder = "TIME",           -- "DEFAULT" / "TIME" / "NAME"
 
@@ -2705,7 +2703,6 @@ DF.RaidDefaults = {
     directBuffFilterRaidInCombat = true,      -- RAID_IN_COMBAT filter
     directBuffFilterCancelable = false,       -- CANCELABLE filter
     directBuffFilterNotCancelable = false,    -- NOT_CANCELABLE filter
-    directBuffFilterImportant = true,         -- IMPORTANT filter (12.0.1)
     directBuffFilterBigDefensive = true,      -- BIG_DEFENSIVE filter (12.0.1)
     directBuffFilterExternalDefensive = true, -- EXTERNAL_DEFENSIVE filter (12.0.0)
     directBuffSortOrder = "TIME",             -- "DEFAULT" / "TIME" / "NAME"
@@ -2715,7 +2712,6 @@ DF.RaidDefaults = {
     directDebuffFilterRaid = true,            -- RAID filter
     directDebuffFilterRaidInCombat = true,    -- RAID_IN_COMBAT filter
     directDebuffFilterCrowdControl = true,    -- CROWD_CONTROL filter
-    directDebuffFilterImportant = true,       -- IMPORTANT filter (12.0.1)
     directDebuffDispellableMode = "PLAYER",  -- "PLAYER" (RAID_PLAYER_DISPELLABLE) or "ALL" (dispelName ~= nil)
     directDebuffSortOrder = "TIME",           -- "DEFAULT" / "TIME" / "NAME"